Null-terminate pam responses in the converse function
The username and password response was previously directly created from
rust's str, which is not null-terminated. Depending on the allocator used,
this usually led to a PAM_USER_UNKNOWN error, because the username returned
contained additional characters at the end.

The fix is to properly null-terminate the strings using CString before
cloning them with strdup for the response.